version2/app/requirements.txt
psycopg2-binary
pandas
numpy
pyarrow